+++ docs/changelog.md
@@ -13,6 +13,10 @@ full list of changes to the project, please consult the commit log.
 ### Changed
 - Further explained details of the Extra Chance changes in the 1.5.0
   changelog.
+- Ancient Blood Stone:
+    - Reduced base chance to cast Blood Lightning from `40%` to `20%`.
+    - Each additional Blood Stone or Ancient Blood Stone increases the
+      chance to cast Blood Lightning by `10%`.
 
 ### Fixed
 - Handling of rank related to players Green and Pink reaching the Damage